chore: restructure directory
